langchain/vector_embed_local_chroma_client.py
import os
from dotenv import find_dotenv, load_dotenv
import openai
import chromadb
from chromadb.utils import embedding_functions
from langchain.chat_models import ChatOpenAI
from langchain.llms import OpenAI
from langchain.chains import LLMChain
from langchain.prompts import PromptTemplate
from langchain_openai import OpenAIEmbeddings
from langchain import LLMChain, PromptTemplate
from langchain.vectorstores import Chroma
from langchain.text_splitter import CharacterTextSplitter, RecursiveCharacterTextSplitter
import numpy as np

load_dotenv()
OPENAI_API_KEY = os.getenv("OPENAI_API_KEY")
llm_model = "gpt-4o-mini"
chat = ChatOpenAI(temperature = 0.0, model = llm_model, openai_api_key=OPENAI_API_KEY)
embeddings = OpenAIEmbeddings()

client = chromadb.HttpClient(host="localhost", port=8000)

openai_ef = embedding_functions.OpenAIEmbeddingFunction(
    api_key=OPENAI_API_KEY,
    model_name="text-embedding-3-small"
)

collection = client.get_or_create_collection(
    name="docs_collection",
    embedding_function=openai_ef
)

results = collection.query(
    query_texts=["My hope is for freedom to ring across every town and every state"],
    n_results=5
)

print("Query results:", results)